More (int) casts to reduce warnings.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2342 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/HyPerCol.cpp
@@ -160,8 +160,8 @@ int HyPerCol::commRow(int colId)
 int HyPerCol::setLayerLoc(PVLayerLoc * layerLoc,
                           float nxScale, float nyScale, int margin, int nf)
 {
-   layerLoc->nxGlobal = nxScale * imageLoc.nxGlobal;
-   layerLoc->nyGlobal = nyScale * imageLoc.nyGlobal;
+   layerLoc->nxGlobal = (int) (nxScale * imageLoc.nxGlobal);
+   layerLoc->nyGlobal = (int) (nyScale * imageLoc.nyGlobal);
 
    // partition input space based on the number of processor
    // columns and rows
